refactor HeroContent and ContactForm to new components
